Adding the graphs I made
The titles are pretty self-explanatory but it should be noted that the combined doesn't show the runtimes for 100k points because I wanted to be able to see more detail. Also, I have yet to make the graphs for the optimized code. When I do, I'll recreate these three files, plus files comparing the optimized code with the un-optimized code.